requirements.txt
blinker==1.5
click==8.1.3
Flask==2.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-decouple==3.6
python-dotenv==0.20.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
zipp==3.8.0